A tutorial on how to use some string functions
The string functions used in this tutorial are the substring, whos
parameters are the index of where you want to start, and where you want
to end. It returns the length of the string from those two points. The
second string function is called the swap. This one simply swaps a
string with another string. Lastly, the tutorial teaches the find
function. It locates where a desired string takes place. Use rfind to
search from right to left.